add old review changes
